backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1

# PDF & NLP
PyMuPDF==1.24.9
textstat==0.7.4
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.2
numpy==1.26.4
nltk==3.9.1

# Optional LLM (LangChain + OpenAI-compatible)
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
tiktoken==0.7.0
